Fix bbannlib issue

+++ include/util/constants.h
@@ -35,7 +35,7 @@ constexpr static int MAX_SAME_SIZE_THRESHOLD = 1500;
 // the prunning rate of dynamic search
 constexpr static float SEARCH_PRUNING_RATE = 0.9;
 
-constexpr static float HNSW_BUCKET_SAMPLE = 4;
+constexpr static int HNSW_BUCKET_SAMPLE = 4;
 
 // file prefix strings and suffix strings
 // file name rule: prefix + file_name + file_type + suffix
